move TimeSettings to common module
