+++ nomos-services/blend/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod core;
 pub mod edge;
 pub mod message;
+pub mod settings;
